requirements.txt
black>=21.9b0
eth-brownie>=1.16.0
dotmap==1.3.23
python-dotenv==0.16.0
tabulate==0.8.7
rich==11.2.0
